+++ changelog.md
@@ -14,6 +14,7 @@ Codion Change Log
 - Resources.get() renamed getString(), consistent with ResourceBundle.
 ### is codion.common.db
 - Database.databaseException() renamed exception().
+- Database.createConnectionPool() now returns the pool.
 ### is codion.common.model
 - ColumnSummaryModel.SummaryValues removed, related refactoring.
 - ColumnSummaryModel.SummaryValueProvider renamed SummaryValues, related renaming.